requirements.txt
tensorflow>=2.0
gin-config
